transport: rdp: let use space close connection
The CSP router task usually runs with high priority to process incoming
packets as fast as possible. When using RDP for short lived connections,
the stack could go through the full transmission and RST sequency before
the client actually got a chance to process the incoming data. When the
RDP implementation calls csp_close(), it would clear the incoming data
queue.

This patch changes the logic to post a NULL pointer on the connection
when closing, and let the client close when it has processed all queued
messages. Unfortunately, the csp_read interface does not have a method
to distinguish between timeouts and connection closed, so returning NULL
is our only option here.